requirements.txt
streamlit
google-generativeai
audiorecorder
SpeechRecognition
gTTS
pydub